Split up extension into KA part and CUDA part. Also restore CPU error messages.
