+++ Project.toml
@@ -26,6 +26,7 @@ SparseMatricesCSR = "a0a7dd2c-ebf4-11e9-1f05-cf50bc540ca1"
 [extensions]
 FerriteBlockArrays = "BlockArrays"
 FerriteCudaExt = ["Adapt", "CUDA", "KernelAbstractions"]
+FerriteKAExt = ["Adapt", "KernelAbstractions"]
 FerriteMetis = "Metis"
 FerriteSparseMatrixCSR = "SparseMatricesCSR"
